+++ src/Guilded.Commands/attributes/CommandParamAttribute.cs
@@ -1,10 +1,35 @@
 using System;
+using Guilded.Base;
 
 namespace Guilded.Commands;
 
 /// <summary>
 /// Declares a parameter as a <see cref="CommandAttribute">command's</see> parameter.
 /// </summary>
+/// <remarks>
+/// <para>Currently available parameter types are:</para>
+/// <list type="bullet">
+///     <item><see cref="string" /></item>
+///     <item><see cref="HashId" /></item>
+///     <item><see cref="Guid" /></item>
+///     <item><see cref="char" /></item>
+///     <item><see cref="bool" /></item>
+///     <item><see cref="int" /></item>
+///     <item><see cref="uint" /></item>
+///     <item><see cref="long" /></item>
+///     <item><see cref="ulong" /></item>
+///     <item><see cref="short" /></item>
+///     <item><see cref="ushort" /></item>
+///     <item><see cref="byte" /></item>
+///     <item><see cref="sbyte" /></item>
+///     <item><see cref="float" /></item>
+///     <item><see cref="double" /></item>
+///     <item><see cref="decimal" /></item>
+///     <item><see cref="DateTime" /></item>
+///     <item><see cref="TimeSpan" /></item>
+/// </list>
+/// <para><see cref="string" /> <see cref="Array" /> is also available, but it will be seen as a rest argument.</para>
+/// </remarks>
 /// <example>
 /// <para>Here's an example of a parameter being declared as a command parameter:</para>
 /// <code language="csharp">